+++ packages/graph/src/core/helpers/edge.ts
