+++ packages/graph/src/core/helpers/index.ts
@@ -1,12 +1,13 @@
-import { CoreEventMap } from '../core/events.ts';
-import { CoreGraph } from '../core/types.ts';
 import { edgeHelpers } from './edge.ts';
 import { nodeHelpers } from './node.ts';
-import { EdgeHelpers, GraphHelpers, NodeHelpers } from './types.ts';
+import {
+  CoreGraphForHelpers,
+  EdgeHelpers,
+  GraphHelpers,
+  NodeHelpers,
+} from './types.ts';
 
-export const useGraphHelpers = <A, B extends CoreEventMap, C>(
-  graph: CoreGraph<A, B, C>,
-): GraphHelpers => ({
+export const useGraphHelpers = (graph: CoreGraphForHelpers): GraphHelpers => ({
   nodes: Object.keys(nodeHelpers).reduce(
     (acc, curr) => ({ ...acc, [curr]: (nodeHelpers as any)[curr](graph) }),
     {} as NodeHelpers,
